+++ acceptance-tests/src/acceptance-test/java/tech/pegasys/teku/test/acceptance/validatorslashing/MultiPeersSingleProcessBlocksAcceptanceTest.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.test.acceptance.slasshingdetection;
+package tech.pegasys.teku.test.acceptance.validatorslashing;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
@@ -24,18 +24,18 @@ import tech.pegasys.teku.test.acceptance.dsl.TekuNode;
  * Running 2 nodes with VC/BN running in single processes. <br>
  * The slashing event is sent to the first node via the POST attester/proposer slashing REST API.
  * It's then <br>
- * received by the second node (running the slashed validator) through the attester/proposer
- * slashings gossip topic <br>
- * (The first node is not running any validator and hence won't produce any block to include the
- * slashing event)
+ * received by the second node (running the slashed validator) within a block (The slashing event is
+ * sent <br>
+ * to the first node before starting the second one. The second node receives the slashing event
+ * within <br>
+ * the block when syncing)
  */
-public class MultiPeersSingleProcessGossipAcceptanceTest
+public class MultiPeersSingleProcessBlocksAcceptanceTest
     extends ValidatorSlashingDetectionAcceptanceTest {
-
   @ParameterizedTest
   @MethodSource("getSlashingEventTypes")
   void
-      shouldShutDownWhenOwnedValidatorSlashed_SingleProcess_MultiplePeers_SlashingEventsThroughGossipOnly_NoBlocks(
+      shouldShutDownWhenOwnedValidatorSlashed_SingleProcess_MultiplePeers_SlashingThroughBlock_NoSlashingEventsGossip(
           final SlashingEventType slashingEventType) throws Exception {
 
     final int genesisTime = timeProvider.getTimeInSeconds().plus(10).intValue();
@@ -46,26 +46,14 @@ public class MultiPeersSingleProcessGossipAcceptanceTest
             config ->
                 configureNode(config, genesisTime, network)
                     .withRealNetwork()
-                    .withAltairEpoch(altairEpoch));
+                    .withAltairEpoch(altairEpoch)
+                    .withInteropValidators(0, 32));
 
     firstTekuNode.start();
 
-    firstTekuNode.waitForEpochAtOrAbove(1);
-
-    final TekuNode secondTekuNode =
-        createTekuNode(
-            config ->
-                configureNode(config, genesisTime, network)
-                    .withAltairEpoch(altairEpoch)
-                    .withStopVcWhenValidatorSlashedEnabled()
-                    .withInteropValidators(0, 32)
-                    .withPeers(firstTekuNode));
-
-    secondTekuNode.start();
-
     firstTekuNode.waitForEpochAtOrAbove(2);
 
-    final int slashedValidatorIndex = 4;
+    final int slashedValidatorIndex = 34;
     final BLSKeyPair slashedValidatorKeyPair = getBlsKeyPair(slashedValidatorIndex);
     final int slotInFirstEpoch =
         firstTekuNode.getSpec().forMilestone(SpecMilestone.ALTAIR).getSlotsPerEpoch() - 1;
@@ -76,6 +64,17 @@ public class MultiPeersSingleProcessGossipAcceptanceTest
         UInt64.valueOf(slashedValidatorIndex),
         slashedValidatorKeyPair.getSecretKey(),
         slashingEventType);
+
+    final TekuNode secondTekuNode =
+        createTekuNode(
+            config ->
+                configureNode(config, genesisTime, network)
+                    .withAltairEpoch(altairEpoch)
+                    .withStopVcWhenValidatorSlashedEnabled()
+                    .withInteropValidators(32, 32)
+                    .withPeers(firstTekuNode));
+
+    secondTekuNode.start();
 
     secondTekuNode.waitForLogMessageContaining(
         String.format(slashingActionLog, slashedValidatorKeyPair.getPublicKey().toHexString()));
