+++ acceptance-tests/src/acceptance-test/java/tech/pegasys/teku/test/acceptance/validatorslashing/MultiPeersStandAloneVcGossipAcceptanceTest.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.test.acceptance.slasshingdetection;
+package tech.pegasys.teku.test.acceptance.validatorslashing;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
@@ -27,15 +27,16 @@ import tech.pegasys.teku.test.acceptance.dsl.TekuValidatorNode;
  * - Node 2: Stand-alone VC with a separate BN <br>
  * The slashing event is sent to the first node via the POST attester/proposer slashing REST API. It
  * is then sent <br>
- * to the second BN withing a block which sends it to it's VC via the attester/proposer slashing SSE
- * channel
+ * to the second BN via the attester/proposer slashing gossip topic which sends it to it's VC via
+ * the <br>
+ * attester/proposer slashing SSE channel
  */
-public class MultiPeersStandAloneVcBlocksAcceptanceTest
+public class MultiPeersStandAloneVcGossipAcceptanceTest
     extends ValidatorSlashingDetectionAcceptanceTest {
   @ParameterizedTest
   @MethodSource("getSlashingEventTypes")
   void
-      shouldShutDownWhenOwnedValidatorSlashed_StandAloneVC_MultiplePeers_SlashingThroughBlock_NoSlashingEventsGossip(
+      shouldShutDownWhenOwnedValidatorSlashed_StandAloneVC_MultiplePeers_SlashingEventsThroughGossipOnly_NoBlocks(
           final SlashingEventType slashingEventType) throws Exception {
 
     final int genesisTime = timeProvider.getTimeInSeconds().plus(10).intValue();
@@ -46,24 +47,11 @@ public class MultiPeersStandAloneVcBlocksAcceptanceTest
             config ->
                 configureNode(config, genesisTime, network)
                     .withRealNetwork()
-                    .withAltairEpoch(altairEpoch)
-                    .withInteropValidators(0, 32));
+                    .withAltairEpoch(altairEpoch));
 
     firstTekuNode.start();
 
-    firstTekuNode.waitForEpochAtOrAbove(2);
-
-    final int slashedValidatorIndex = 34;
-    final BLSKeyPair slashedValidatorKeyPair = getBlsKeyPair(slashedValidatorIndex);
-    final int slotInThirdEpoch =
-        firstTekuNode.getSpec().forMilestone(SpecMilestone.ALTAIR).getSlotsPerEpoch() * 2 + 3;
-
-    postSlashing(
-        firstTekuNode,
-        UInt64.valueOf(slotInThirdEpoch),
-        UInt64.valueOf(slashedValidatorIndex),
-        slashedValidatorKeyPair.getSecretKey(),
-        slashingEventType);
+    firstTekuNode.waitForEpochAtOrAbove(1);
 
     final TekuNode secondBeaconNode =
         createTekuNode(
@@ -80,12 +68,26 @@ public class MultiPeersStandAloneVcBlocksAcceptanceTest
                     .withNetwork("auto")
                     .withValidatorApiEnabled()
                     .withStopVcWhenValidatorSlashedEnabled()
-                    .withInteropValidators(32, 32)
+                    .withInteropValidators(0, 32)
                     .withBeaconNode(secondBeaconNode));
 
     secondBeaconNode.start();
 
     secondValidatorClient.start();
+
+    firstTekuNode.waitForEpochAtOrAbove(2);
+
+    final int slashedValidatorIndex = 4;
+    final BLSKeyPair slashedValidatorKeyPair = getBlsKeyPair(slashedValidatorIndex);
+    final int slotInThirdEpoch =
+        firstTekuNode.getSpec().forMilestone(SpecMilestone.ALTAIR).getSlotsPerEpoch() * 2 + 3;
+
+    postSlashing(
+        firstTekuNode,
+        UInt64.valueOf(slotInThirdEpoch),
+        UInt64.valueOf(slashedValidatorIndex),
+        slashedValidatorKeyPair.getSecretKey(),
+        slashingEventType);
 
     secondValidatorClient.waitForLogMessageContaining(
         String.format(slashingActionLog, slashedValidatorKeyPair.getPublicKey().toHexString()));
@@ -94,9 +96,6 @@ public class MultiPeersStandAloneVcBlocksAcceptanceTest
 
     // Make sure the BN didn't shut down
     secondBeaconNode.waitForBlockAtOrAfterSlot(4);
-    // Make sure the first node didn't shut down
-    firstTekuNode.waitForBlockAtOrAfterSlot(4);
     secondBeaconNode.stop();
-    firstTekuNode.stop();
   }
 }
