+++ acceptance-tests/src/acceptance-test/java/tech/pegasys/teku/test/acceptance/validatorslashing/SinglePeerAcceptanceTest.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.test.acceptance.slasshingdetection;
+package tech.pegasys.teku.test.acceptance.validatorslashing;
 
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
@@ -19,57 +19,48 @@ import tech.pegasys.teku.bls.BLSKeyPair;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.spec.SpecMilestone;
 import tech.pegasys.teku.test.acceptance.dsl.TekuNode;
-import tech.pegasys.teku.test.acceptance.dsl.TekuValidatorNode;
 
 /**
- * Running a stand-alone VC with a separate BN. The slashing event is sent to the BN via the POST
- * attester/proposer slashing POST API
+ * Running a single node with BN/VC running in a single process. The slashing event is sent to the
+ * node via the POST attester/proposer slashing POST API
  */
-public class SinglePeerStandAloneVcAcceptanceTest extends ValidatorSlashingDetectionAcceptanceTest {
+public class SinglePeerAcceptanceTest extends ValidatorSlashingDetectionAcceptanceTest {
 
   @ParameterizedTest
   @MethodSource("getSlashingEventTypes")
-  void shouldShutDownWhenOwnedValidatorSlashed_StandAloneVC_SinglePeer(
-      final SlashingEventType slashingEventType) throws Exception {
+  void shouldShutDownWhenOwnedValidatorSlashed_SingleProcess_SinglePeer(
+      final ValidatorSlashingDetectionAcceptanceTest.SlashingEventType slashingEventType)
+      throws Exception {
 
     final int genesisTime = timeProvider.getTimeInSeconds().plus(10).intValue();
     final UInt64 altairEpoch = UInt64.valueOf(100);
 
-    final TekuNode beaconNode =
+    final TekuNode tekuNode =
         createTekuNode(
-            config -> configureNode(config, genesisTime, network).withAltairEpoch(altairEpoch));
-
-    final TekuValidatorNode validatorClient =
-        createValidatorNode(
             config ->
-                config
-                    .withNetwork("auto")
+                configureNode(config, genesisTime, network)
+                    .withAltairEpoch(altairEpoch)
                     .withStopVcWhenValidatorSlashedEnabled()
-                    .withInteropValidators(0, 32)
-                    .withBeaconNode(beaconNode));
+                    .withInteropValidators(0, 32));
 
-    beaconNode.start();
-    validatorClient.start();
+    tekuNode.start();
 
-    beaconNode.waitForEpochAtOrAbove(2);
+    tekuNode.waitForEpochAtOrAbove(2);
 
     final int slashedValidatorIndex = 3;
     final BLSKeyPair slashedValidatorKeyPair = getBlsKeyPair(slashedValidatorIndex);
     final int slotInFirstEpoch =
-        beaconNode.getSpec().forMilestone(SpecMilestone.ALTAIR).getSlotsPerEpoch() - 1;
+        tekuNode.getSpec().forMilestone(SpecMilestone.ALTAIR).getSlotsPerEpoch() - 1;
 
     postSlashing(
-        beaconNode,
+        tekuNode,
         UInt64.valueOf(slotInFirstEpoch),
         UInt64.valueOf(slashedValidatorIndex),
         slashedValidatorKeyPair.getSecretKey(),
         slashingEventType);
 
-    validatorClient.waitForLogMessageContaining(
+    tekuNode.waitForLogMessageContaining(
         String.format(slashingActionLog, slashedValidatorKeyPair.getPublicKey().toHexString()));
-    validatorClient.waitForExit(shutdownWaitingSeconds);
-    // Make sure the BN didn't shut down
-    beaconNode.waitForEpochAtOrAbove(4);
-    beaconNode.stop();
+    tekuNode.waitForExit(shutdownWaitingSeconds);
   }
 }
